requirements.txt
firebase-admin
python-dotenv
serpapi
google-generativeai